requirements.txt
fastapi
pandas
numpy
pyarrow
python-dotenv
scikit-learn
requests
pydantic
sentence-transformers
faiss-cpu
rank-bm25
openai==1.95.1
